DF-6961 get exitcode of process for win platform
Former-commit-id: ed0b01984a796497ddf9695260019e78871a5409

+++ Sources/Internal/Platform/Process.h
@@ -54,7 +54,7 @@ public:
     const FilePath& GetPath() const;
     const Vector<String>& GetArgs() const;
 
-    int GetExitCode() const;
+    long GetExitCode() const;
     
 private:
